Support for scheduled events with same names or date/time rules
Fixes #1061
