+++ Common/Scheduling/ScheduleManager.cs
@@ -1,11 +1,11 @@
 ﻿/*
  * QUANTCONNECT.COM - Democratizing Finance, Empowering Individuals.
  * Lean Algorithmic Trading Engine v2.0. Copyright 2014 QuantConnect Corporation.
- * 
- * Licensed under the Apache License, Version 2.0 (the "License"); 
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -177,12 +177,12 @@ namespace QuantConnect.Scheduling
             if (exampleTimes.Length > 0)
             {
                 Log.Trace("Event Name \"{0}\", scheduled to run at {1} (UTC){2}",
-                        name, string.Join(", ", exampleTimes),
+                        scheduledEvent.Name, string.Join(", ", exampleTimes),
                         exampleTimes.Length > 1? "..." : "");
             }
             else
             {
-                Log.Trace("Event Name \"{0}\", scheduled to run, but no event times were selected", name);
+                Log.Trace("Event Name \"{0}\", scheduled to run, but no event times were selected", scheduledEvent.Name);
             }
 
             return scheduledEvent;
